django/requirements.txt
anyio==4.0.0
asgiref==3.7.2
channels==4.0.0
click==8.1.7
colorama==0.4.6
Django==4.2.5
djangorestframework==3.14.0
exceptiongroup==1.1.3
h11==0.14.0
httptools==0.7.1
idna==3.4
psycopg2-binary==2.9.11
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
redis==5.0.5
django-redis==5.4.0